+++ Sources/MapboxCoreNavigation/NavigationSettings.swift
@@ -14,7 +14,7 @@ public typealias RoutingProviderSource = MapboxRoutingProvider.Source
  To customize the user experience during a particular turn-by-turn navigation session, use the `NavigationOptions` class
  when initializing a `NavigationViewController`.
 
- To customize some global defaults use `NavigationSettings.initialize(directions:tileStoreConfiguration:routingProviderSource:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+ To customize some global defaults use `NavigationSettings.initialize(directions:tileStoreConfiguration:routingProviderSource:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
  */
 public class NavigationSettings {
     
@@ -40,7 +40,8 @@ public class NavigationSettings {
                   routingProviderSource: .hybrid,
                   alternativeRouteDetectionStrategy: .init(),
                   utilizeSensorData: false,
-                  navigatorPredictionInterval: nil)
+                  navigatorPredictionInterval: nil,
+                  liveIncidentsOptions: nil)
         }
 
         var directions: Directions
@@ -49,6 +50,7 @@ public class NavigationSettings {
         var alternativeRouteDetectionStrategy: AlternativeRouteDetectionStrategy?
         var utilizeSensorData: Bool
         var navigatorPredictionInterval: TimeInterval?
+        var liveIncidentsOptions: IncidentsOptions?
     }
 
     /// Protects access to `_state`.
@@ -72,7 +74,7 @@ public class NavigationSettings {
     /**
      Default `Directions` instance. By default, `Directions.shared` is used.
 
-     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
      */
     public var directions: Directions {
         state.directions
@@ -81,14 +83,14 @@ public class NavigationSettings {
     /**
      Global `TileStoreConfiguration` instance.
 
-     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
      */
     public var tileStoreConfiguration: TileStoreConfiguration {
         state.tileStoreConfiguration
     }
 
     /**
-     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
      */
     public var routingProviderSource: RoutingProviderSource {
         state.routingProviderSource
@@ -97,7 +99,7 @@ public class NavigationSettings {
     /**
      Configuration on how `AlternativeRoute`s will be detected during navigation process.
      
-     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
      
      If set to `nil`, the detection is turned off.
      */
@@ -123,10 +125,21 @@ public class NavigationSettings {
      
      If not specified (`nilled`), default value will be used.
      
-     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:)` method.
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
      */
     public var navigatorPredictionInterval: TimeInterval? {
         state.navigatorPredictionInterval
+    }
+    
+    /**
+     Configuration on how live incidents on a most probable path are detected.
+     
+     You can override this property by using `NavigationSettings.initialize(directions:tileStoreConfiguration:navigationRouterType:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:)` method.
+     
+     If set to `nil`, live incidents are turned off (by default).
+     */
+    public var liveIncidentsOptions: IncidentsOptions? {
+        state.liveIncidentsOptions
     }
     
     /**
@@ -146,25 +159,28 @@ public class NavigationSettings {
        - alternativeRouteDetectionStrategy: Configures how `AlternativeRoute`s will be detected during navigation process.
        - utilizeSensorData: Enables using sensors data to improve positioning.
        - navigatorPredictionInterval: Defines approximate navigator prediction between location ticks.
+       - liveIncidentsOptions: Configures Electronic Horizon live incidents.
      */
     public func initialize(directions: Directions,
                            tileStoreConfiguration: TileStoreConfiguration,
                            routingProviderSource: RoutingProviderSource = .hybrid,
                            alternativeRouteDetectionStrategy: AlternativeRouteDetectionStrategy? = .init(),
                            utilizeSensorData: Bool = false,
-                           navigatorPredictionInterval: TimeInterval? = nil) {
+                           navigatorPredictionInterval: TimeInterval? = nil,
+                           liveIncidentsOptions: IncidentsOptions? = nil) {
         lock.lock(); defer {
             lock.unlock()
         }
         if _state != nil {
-            Log.warning("Warning: Using NavigationSettings.initialize(directions:tileStoreConfiguration:routingProviderSource:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:) after corresponding variables was initialized. Possible reasons: Initialize called more than once, or the following properties was accessed before initialization: `tileStoreConfiguration`, `directions`, `routingProviderSource`, `alternativeRouteDetectionStrategy`, `utilizeSensorData`, `navigatorPredictionInterval`. This might result in an undefined behaviour.", category: .settings)
+            Log.warning("Warning: Using NavigationSettings.initialize(directions:tileStoreConfiguration:routingProviderSource:alternativeRouteDetectionStrategy:utilizeSensorData:navigatorPredictionInterval:liveIncidentsOptions:) after corresponding variables was initialized. Possible reasons: Initialize called more than once, or the following properties was accessed before initialization: `tileStoreConfiguration`, `directions`, `routingProviderSource`, `alternativeRouteDetectionStrategy`, `utilizeSensorData`, `navigatorPredictionInterval`, `liveIncidentsOptions`. This might result in an undefined behaviour.", category: .settings)
         }
         _state = .init(directions: directions,
                        tileStoreConfiguration: tileStoreConfiguration,
                        routingProviderSource: routingProviderSource,
                        alternativeRouteDetectionStrategy: alternativeRouteDetectionStrategy,
                        utilizeSensorData: utilizeSensorData,
-                       navigatorPredictionInterval: navigatorPredictionInterval)
+                       navigatorPredictionInterval: navigatorPredictionInterval,
+                       liveIncidentsOptions: liveIncidentsOptions)
     }
     
     /**